modified sele-sort
