+++ AlgorithmVisualizer.cpp
@@ -293,6 +293,21 @@ void mergeSort(int a[], int si, int ei)
     merge2SortedArrays(a, si, ei);
 }
 
+void selectionSort(int arr[], int n)
+{
+    for (int i = 0; i < n-1; i++)
+    {
+        int min_idx = i;
+        for (int j = i+1; j < n; j++)
+          if (arr[j] < arr[min_idx])
+            min_idx = j;
+
+        int temp = arr[min_idx];
+        arr[min_idx] = arr[i];
+        arr[i] = temp;
+    }
+}
+
