--- selection_Sort/selection_Sort.cpp
@@ -1,34 +1,0 @@
-#include <iostream>
-using namespace std;
-
-int swap(int *a, int *b){
-    int temp = *a;
-    *a = *b;
-    *b = temp;
-}
-
-void selectionSort(int arr[], int n)
-{
-    // One by one move boundary of unsorted subarray
-    for (int i = 0; i < n-1; i++)
-    {
-        // Find the minimum element in unsorted array
-        int min_idx = i;
-        for (int j = i+1; j < n; j++)
-          if (arr[j] < arr[min_idx])
-            min_idx = j;
-
-        // Swap the found minimum element with the first element
-        swap(&arr[min_idx], &arr[i]);
-    }
-}
-//Time Complexity: O(N2)
-
-int main(){
-    int arr[]={34,32,89,43,2,10,7,14};
-    int n = sizeof(arr)/sizeof(arr[0]);
-    selectionSort(arr,n);
-    for(int i=0; i<n; i++){
-        cout<<"arr["<<i<<"]  =  "<<arr[i]<<endl;
-    }
-}
